requirements.txt
SQLAlchemy==1.4.1
fastapi>=0.75.0
uvicorn>=0.17.6
pydantic>=1.0.0
gunicorn>=20.1.0
python-dotenv